webnn/model-tests/image-classification/mobilenet_nhwc.https.html

<!doctype html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>test MobileNet V2 model with NHWC layout</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/libs/numpy.js"></script>
<script src="../resources/utils.js"></script>
<script src="../../resources/utils.js"></script>
<script>
let context;
let builder;
let graph;

const weightsUrl = '../resources/test-data/models/mobilenetv2_nhwc/weights/';
const outputDimensions = [1, 1001];

const buildConv = async (
    input, weightsSubName, biasSubName, relu6, options) => {
  const weightsName = weightsUrl + 'Const_' + weightsSubName + '.npy';
  const weights = await buildConstantByNpy(builder, weightsName);
  const biasName = weightsUrl + 'MobilenetV2_' + biasSubName + '_bias.npy';
  const bias = await buildConstantByNpy(builder, biasName);
  options.inputLayout = 'nhwc';
  options.bias = bias;
  if (relu6) {
    // `relu6` in TFLite equals to `clamp` in WebNN API
    options.activation = builder.clamp({minValue: 0, maxValue: 6});
  } else {
    options.activation = undefined;
  }
  return builder.conv2d(input, weights, options);
}

const buildLinearBottleneck = async (
    input, weightsNameArray, biasName, dwiseOptions, shortcut = true) => {
  const autoPad = 'same-upper';
  const biasPrefix = 'expanded_conv_' + biasName;

  dwiseOptions.autoPad = autoPad;
  dwiseOptions.filterLayout = 'ihwo';
  const convOptions = {autoPad, filterLayout: 'ohwi'};

  const conv1x1Relu6 = await buildConv(
      input, weightsNameArray[0], `${biasPrefix}_expand_Conv2D`, true,
      convOptions);
  const dwise3x3Relu6 = await buildConv(
      conv1x1Relu6, weightsNameArray[1], `${biasPrefix}_depthwise_depthwise`,
      true, dwiseOptions);
  const conv1x1Linear = await buildConv(
      dwise3x3Relu6, weightsNameArray[2], `${biasPrefix}_project_Conv2D`,
      false, convOptions);

  if (shortcut) {
    return builder.add(input, conv1x1Linear);
  }
  return conv1x1Linear;
}

const load = async () => {
  const strides = [2, 2];
  const autoPad = 'same-upper';
  const filterLayout = 'ohwi';
  const input = builder.input(
      'input', {type: 'float32', dimensions: [1, 224, 224, 3]});
  const conv0 = await buildConv(
      input, '90', 'Conv_Conv2D', true, {strides, autoPad, filterLayout});
  const conv1 = await buildConv(
      conv0, '238', 'expanded_conv_depthwise_depthwise', true,
      {autoPad, groups: 32, filterLayout: 'ihwo'});
  const conv2 = await buildConv(
      conv1, '167', 'expanded_conv_project_Conv2D', false,
      {autoPad, filterLayout});
  const bottleneck0 = await buildLinearBottleneck(
      conv2, ['165', '99', '73'], '1', {strides, groups: 96}, false);
  const bottleneck1 = await buildLinearBottleneck(
      bottleneck0, ['3', '119', '115'], '2', {groups: 144});
  const bottleneck2 = await buildLinearBottleneck(
      bottleneck1, ['255', '216', '157'], '3', {strides, groups: 144}, false);
  const bottleneck3 = await buildLinearBottleneck(
      bottleneck2, ['227', '221', '193'], '4', {groups: 192});
  const bottleneck4 = await buildLinearBottleneck(
      bottleneck3, ['243', '102', '215'], '5', {groups: 192});
  const bottleneck5 = await buildLinearBottleneck(
      bottleneck4, ['226', '163', '229'], '6', {strides, groups: 192}, false);
  const bottleneck6 = await buildLinearBottleneck(
      bottleneck5, ['104', '254', '143'], '7', {groups: 384});
  const bottleneck7 = await buildLinearBottleneck(
      bottleneck6, ['25', '142', '202'], '8', {groups: 384});
  const bottleneck8 = await buildLinearBottleneck(
      bottleneck7, ['225', '129', '98'], '9', {groups: 384});
  const bottleneck9 = await buildLinearBottleneck(
      bottleneck8, ['169', '2', '246'], '10', {groups: 384}, false);
  const bottleneck10 = await buildLinearBottleneck(
      bottleneck9, ['162', '87', '106'], '11', {groups: 576});
  const bottleneck11 = await buildLinearBottleneck(
      bottleneck10, ['52', '22', '40'], '12', {groups: 576});
  const bottleneck12 = await buildLinearBottleneck(
      bottleneck11, ['114', '65', '242'], '13', {strides, groups: 576}, false);
  const bottleneck13 = await buildLinearBottleneck(
      bottleneck12, ['203', '250', '92'], '14', {groups: 960});
  const bottleneck14 = await buildLinearBottleneck(
      bottleneck13, ['133', '130', '258'], '15', {groups: 960});
  const bottleneck15 = await buildLinearBottleneck(
      bottleneck14, ['60', '248', '100'], '16', {groups: 960}, false);
  const conv3 = await buildConv(
      bottleneck15, '71', 'Conv_1_Conv2D', true, {autoPad, filterLayout});

  const averagePool2d = builder.averagePool2d(
      conv3, {windowDimensions: [7, 7], layout: 'nhwc'});
  const conv4 = await buildConv(
      averagePool2d, '222', 'Logits_Conv2d_1c_1x1_Conv2D', false,
      {autoPad, filterLayout});
  const reshape = builder.reshape(conv4, [1, null]);
  return builder.softmax(reshape);
}

const build = async (outputOperand) => {
  graph = await builder.build({'output': outputOperand});
}

const compute = async (inputBuffer, outputBuffer) => {
  const inputs = {'input': inputBuffer};
  const outputs = {'output': outputBuffer};
  await context.compute(graph, inputs, outputs);
}

promise_setup(async () => {
  context = await navigator.ml.createContext();
  builder = new MLGraphBuilder(context);
}, {explicit_timeout: true, test_timeout: 180000});

promise_test(async () => {
  const inputElement = await createInputElement('img', './test.jpg');
  const inputBuffer = getInputTensor(inputElement, {
      mean: [127.5, 127.5, 127.5],
      std: [127.5, 127.5, 127.5],
      inputLayout: 'nhwc',
      inputDimensions: [1, 224, 224, 3],
  });
  const outputBuffer = new Float32Array(sizeOfShape(outputDimensions));
  console.time('[total] Load + build graph MobileNet V2 NHWC model');
  const outputOperand = await load();
  graph = await builder.build({'output': outputOperand});
  console.timeEnd('[total] Load + build graph MobileNet V2 NHWC model');
  const inputs = {'input': inputBuffer};
  const outputs = {'output': outputBuffer};
  console.time('Compute with MobileNet V2 NHWC model');
  const results = await context.compute(graph, inputs, outputs);
  console.timeEnd('Compute with MobileNet V2 NHWC model');
  const labels = await fetchLabels('./labels1001.txt');
  const labelClasses = getTopClasses(results.outputs.output, labels);

  assert_equals(labelClasses[0].label, 'lesser panda', 'Get top 1 label OK.');
  assert_equals(labelClasses[1].label, 'polecat', 'Get top 2 label OK');
  assert_equals(labelClasses[2].label, 'giant panda', 'Get top 3 label OK');
}, 'test MobileNet V2 model with NHWC layout');
</script>
